style-transfer/demonstration/pyproject.toml
[project]
name = "demonstrate-neural-style-transfer"
version = "0.1.0"
description = ""
authors = [{ name = "MR", email = "<raffmatth@gmail.com>" }]
requires-python = ">=3.13, <3.14"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
ipykernel = "^6.29"
pillow = "^12.1.1"
numpy = "^2.4.2"
tensorflow = "^2.20.0"
scipy = "^1.17.0"
matplotlib = "^3.10.8"

[[tool.poetry.source]]
name = "PyPI"
priority = "supplemental"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
types-pyyaml = "^6.0.12.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false